requirements.txt
fastapi
uvicorn
langchain-openai
langchain-core
langgraph
python-dotenv